refactor: move light client into dash folder
