+++ dash/light/client_test.go
@@ -15,11 +15,11 @@ import (
 	dbm "github.com/tendermint/tm-db"
 
 	dashcore "github.com/tendermint/tendermint/dash/core"
+	"github.com/tendermint/tendermint/dash/light"
+	"github.com/tendermint/tendermint/dash/light/provider"
+	provider_mocks "github.com/tendermint/tendermint/dash/light/provider/mocks"
+	dbs "github.com/tendermint/tendermint/dash/light/store/db"
 	"github.com/tendermint/tendermint/libs/log"
-	"github.com/tendermint/tendermint/light"
-	"github.com/tendermint/tendermint/light/provider"
-	provider_mocks "github.com/tendermint/tendermint/light/provider/mocks"
-	dbs "github.com/tendermint/tendermint/light/store/db"
 	"github.com/tendermint/tendermint/types"
 )
