+++ dash/light/dash_core_verifier.go
