+++ dash/light/doc.go
