+++ dash/light/errors.go
