+++ dash/light/provider/errors.go
