+++ dash/light/provider/http/http_test.go
@@ -11,8 +11,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/tendermint/tendermint/abci/example/kvstore"
-	"github.com/tendermint/tendermint/light/provider"
-	lighthttp "github.com/tendermint/tendermint/light/provider/http"
+	"github.com/tendermint/tendermint/dash/light/provider"
+	lighthttp "github.com/tendermint/tendermint/dash/light/provider/http"
 	rpcclient "github.com/tendermint/tendermint/rpc/client"
 	rpchttp "github.com/tendermint/tendermint/rpc/client/http"
 	rpcmock "github.com/tendermint/tendermint/rpc/client/mocks"
