+++ dash/light/rpc/client_test.go
