+++ dash/light/store/db/db_test.go
