+++ dash/light/store/errors.go
