+++ dash/light/store/store.go
